Add template chat. Remove image

+++ DShop2024/Areas/Admin/Controllers/BrandManageController.cs
@@ -8,7 +8,7 @@ using Microsoft.EntityFrameworkCore;
 using DShop2024.Models;
 using Microsoft.AspNetCore.Authorization;
 using DShop2024.EnumData;
-using Microsoft.AspNetCore.Hosting;
+
 
 namespace DShop2024.Areas.Admin.Controllers
 {
@@ -74,12 +74,14 @@ namespace DShop2024.Areas.Admin.Controllers
                 {
                     _context.Add(brandModel);
                     await _context.SaveChangesAsync();
+                    TempData[DShopConst.TEMPDATA_SUCCESS] = "Create brand successful ";
                     return RedirectToAction(nameof(Index));
                 }
                 catch (Exception ex)
                 {
 
                     TempData[DShopConst.TEMPDATA_ERROR] = "Create brand fail "+ ex.Message;
+                    return View(brandModel);
                 }
             }
             return View(brandModel);
@@ -93,7 +95,7 @@ namespace DShop2024.Areas.Admin.Controllers
                 return NotFound();
             }
 
-            var brandModel = await _context.Brands.FindAsync(id);
+            var brandModel = await _context.Brands.FirstOrDefaultAsync(m => m.Id == id && m.Status != 0);
             if (brandModel == null)
             {
                 return NotFound();
@@ -132,20 +134,24 @@ namespace DShop2024.Areas.Admin.Controllers
                         string imageName = Guid.NewGuid().ToString() + "_" + brandModel.ImageUpload.FileName;
                         string filePath = Path.Combine(uploadsDir, imageName);
 
-                        string oldFilePath = Path.Combine(uploadsDir, exitedBrand.Image);
-                        try
+                        if(exitedBrand.Image != null)
                         {
-                            if (System.IO.File.Exists(oldFilePath))
+                            string oldFilePath = Path.Combine(uploadsDir, exitedBrand.Image);
+                            try
                             {
-                                System.IO.File.Delete(oldFilePath);
-                            }
+                                if (System.IO.File.Exists(oldFilePath))
+                                {
+                                    System.IO.File.Delete(oldFilePath);
+                                }
 
+                            }
+                            catch (Exception ex)
+                            {
+                                TempData[DShopConst.TEMPDATA_ERROR] = "update brand fail" + ex.Message;
+                                return View(brandModel);
+                            }
                         }
-                        catch (Exception ex)
-                        {
-							TempData[DShopConst.TEMPDATA_ERROR] = "update brand fail"+ ex.Message;
-							return View(brandModel);
-						}
+                     
                         FileStream fs = new FileStream(filePath, FileMode.Create);
                         await brandModel.ImageUpload.CopyToAsync(fs);
                         fs.Close();
@@ -160,8 +166,9 @@ namespace DShop2024.Areas.Admin.Controllers
 
 					_context.Update(exitedBrand);
                     await _context.SaveChangesAsync();
+                    TempData[DShopConst.TEMPDATA_SUCCESS] = "Update brand successful ";
                 }
-                catch (DbUpdateConcurrencyException)
+                catch (DbUpdateConcurrencyException db)
                 {
                     if (!BrandModelExists(brandModel.Id))
                     {
@@ -169,7 +176,8 @@ namespace DShop2024.Areas.Admin.Controllers
                     }
                     else
                     {
-                        throw;
+                        TempData[DShopConst.TEMPDATA_ERROR] = "update brand fail" + db.Message;
+                        return View(brandModel);
                     }
                 }
                 return RedirectToAction(nameof(Index));
@@ -185,22 +193,47 @@ namespace DShop2024.Areas.Admin.Controllers
             {
                 return NotFound();
             }
-
             var brandModel = await _context.Brands
-                .FirstOrDefaultAsync(m => m.Id == id);
+                .FirstOrDefaultAsync(m => m.Id == id && m.Status != 0);
             if (brandModel == null)
             {
                 return NotFound();
             }
-            if (brandModel != null)
+
+            try
             {
+                if (brandModel.Image != null)
+                {
+                    string uploadsDir = Path.Combine(_webHostEnvironment.WebRootPath, "media/brands");
+                    string oldFilePath = Path.Combine(uploadsDir, brandModel.Image);
+                    try
+                    {
+                        if (System.IO.File.Exists(oldFilePath))
+                        {
+                            System.IO.File.Delete(oldFilePath);
+                        }
+
+                    }
+                    catch (Exception ex)
+                    {
+                        TempData[DShopConst.TEMPDATA_ERROR] = "An error occurred while deleting the banner image " + ex.Message;
+                        return RedirectToAction("Index");
+                    }
+                }
+
                 brandModel.Status = 0;
                 _context.Brands.Update(brandModel);
                 await _context.SaveChangesAsync();
+                TempData[DShopConst.TEMPDATA_SUCCESS] = "Delete brand successful";
+                await _context.SaveChangesAsync();
+                return RedirectToAction(nameof(Index));
             }
-			TempData[DShopConst.TEMPDATA_SUCCESS] = "Delete successful";
-			await _context.SaveChangesAsync();
-            return RedirectToAction(nameof(Index));
+            catch (Exception ex)
+            {
+                TempData[DShopConst.TEMPDATA_ERROR] = "Delete brand fail "+ex.Message;
+                return RedirectToAction(nameof(Index));
+            }
+
         }
 
         private bool BrandModelExists(int id)
